requirements.txt
pandas
numpy
scikit-learn
pyyaml
streamlit
pytest

